fix(pipeline): resolve phase 3 review blockers
- Correct the EventBridge schedule role trust policy to use the classic
  `events.amazonaws.com` principal so the Step Functions pipeline triggers
  successfully when the schedule fires.
- Consolidate the SQLite JSONB/UUID @compiles helpers into `tests/conftest.py`
  to remove duplication between the scanner and evaluator unit tests.
- Capture deal IDs within the scanner session context to avoid dereferencing
  detached ORM instances after the session is closed, keeping the agent robust
  against future SQLAlchemy changes.

Validation:
- `uv run pytest tests/ -q`

Part of #4
